+++ src/main/java/kitchenpos/order/enums/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.enums;
+package kitchenpos.order.enums;
 
 public enum OrderStatus {
     COOKING, MEAL, COMPLETION
